+++ index.html
@@ -249,7 +249,7 @@
   </template>
 
 
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 
 </body>
